# docs/swagger.yaml
basePath: /
definitions:
  handler.CreateSubscriptionInput:
    properties:
      end_date:
        example: 12-2023
        type: string
      price:
        example: 599
        type: integer
      service_name:
        example: Netflix
        type: string
      start_date:
        example: 01-2023
        type: string
      user_id:
        example: a1b2c3d4-e5f6-7g8h-9i0j-k1l2m3n4o5p6
        type: string
    type: object
  handler.ErrorResponse:
    properties:
      error:
        example: описание ошибки
        type: string
    type: object
  handler.UpdateSubscriptionInput:
    properties:
      end_date:
        example: 12-2023
        type: string
      price:
        example: 399
        type: integer
      service_name:
        example: Yandex Plus
        type: string
      start_date:
        example: 02-2023
        type: string
    type: object
  model.Subscription:
    properties:
      end_date:
        type: string
      id:
        type: string
      price:
        type: integer
      service_name:
        type: string
      start_date:
        type: string
      user_id:
        type: string
    type: object
host: localhost:8080
info:
  contact: {}
  description: API для управления онлайн-подписками пользователей
  title: Online Subscriptions API
  version: "1.0"
paths:
  /subscriptions:
    get:
      description: Возвращает список всех подписок
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/model.Subscription'
            type: array
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
      summary: Получить все подписки
      tags:
      - subscriptions
    post:
      consumes:
      - application/json
      description: Создаёт новую онлайн-подписку
      parameters:
      - description: Данные подписки
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/handler.CreateSubscriptionInput'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/model.Subscription'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
      summary: Создать подписку
      tags:
      - subscriptions
  /subscriptions/{id}:
    delete:
      description: Удаляет подписку по ID
      parameters:
      - description: UUID подписки
        in: path
        name: id
        required: true
        type: string
      responses:
        "204":
          description: No Content
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
      summary: Удалить подписку
      tags:
      - subscriptions
    put:
      consumes:
      - application/json
      description: Обновляет поля существующей подписки
      parameters:
      - description: UUID подписки
        in: path
        name: id
        required: true
        type: string
      - description: Данные для обновления
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/handler.UpdateSubscriptionInput'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.Subscription'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
      summary: Обновить подписку
      tags:
      - subscriptions
  /subscriptions/{user_id}:
    get:
      description: Возвращает список подписок определённого пользователя
      parameters:
      - description: UUID пользователя
        in: path
        name: user_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/model.Subscription'
            type: array
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
      summary: Получить подписки по user_id
      tags:
      - subscriptions
  /subscriptions/summary:
    get:
      description: Выводит общую сумму подписок по фильтрам
      parameters:
      - description: UUID пользователя
        in: query
        name: user_id
        type: string
      - description: Название сервиса
        in: query
        name: service_name
        type: string
      - description: Начало периода (MM-YYYY)
        in: query
        name: start_date
        required: true
        type: string
      - description: Конец периода (MM-YYYY)
        in: query
        name: end_date
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: integer
            type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
      summary: Получить сумму подписок
      tags:
      - subscriptions
swagger: "2.0"